fix: UI polish for sale-readiness — no model names exposed, settings dropdown crash, test suite green
- Remove model names from UI (settings, wizard, about) via modelDisplayLabel
- Fix DropdownButton assertion: default model tiny → base, dropdown items base/turbo-q5
- Fix RenderFlex overflow in settings tiles (Flexible trailing + isExpanded dropdown)
- Delete duplicate privacy/usage screens in widgets/ (stale copies, leaked model names)
- Wire side panel to canonical screens/ versions
- Fix flutter test: privacy_proof whitelist, settings labels, scrollUntilVisible, wizard detectSpecs injection
- Regenerate FRB bindings (export_sanitize_filename exposed via api.rs)
- CI: remove invalid rust-cache inputs (sync-club, allow-turbo-fetch)
- 55 tests pass, 2 skipped; flutter analyze clean; cargo clippy 0 warnings

+++ rust_core/src/audio/loopback.rs
@@ -132,7 +132,7 @@ pub(crate) mod macos {
                     if let Some(list) = sample.audio_buffer_list() {
                         for buf in list.iter() {
                             let ptr = buf.data().as_ptr() as *const f32;
-                            let len = buf.data_byte_size() as usize / 4;
+                            let len = buf.data_byte_size() / 4;
                             if !ptr.is_null() && len > 0 {
                                 let samples = unsafe { std::slice::from_raw_parts(ptr, len) };
                                 let _ = tx.send(samples.to_vec());
